* improved udhcp rocknet module to support supervision
git-svn-id: https://svn.exactcode.de/t2/trunk@6701 c5f82cb5-29bc-0310-9cd0-bff59a50e3bc

+++ package/network/udhcp/rocknet_udhcp.sh
@@ -1,7 +1,16 @@
 
 public_udhcp() {
-	addcode up   5 5 "/usr/sbin/udhcpc -H `hostname` -i $if"
-	addcode down 5 5 "killall -TERM udhcpc"
-	addcode down 5 6 "sleep 2 ; ip link set $if down || ifconfig $if down"
+	local cmdline="/usr/sbin/udhcpc -h `hostname` -i $if"
+	cmdline="$cmdline -s /etc/udhcp/t2-default.script"
+
+	if [ "$CANUSESERVICE" == "1" ]; then
+		addcode up 5 1 "service_create $if '$cmdline -f'" \
+			"sleep 2 ; ip link set $if down"
+		addcode down 5 1 "service_destroy $if"
+	else
+		addcode up   5 5 "$cmdline"
+		addcode down 5 5 "killall -TERM udhcpc"
+		addcode down 5 5 "sleep 2 ; ip link set $if down"
+	fi
 }
